In  preview_get_template(), Use cms_join_path() instead of hardcoded / for path to preview file. If the template file does not exist give a nice error instead of having php errors occur.
git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/trunk@3764 3d254a34-79dc-0310-9e5f-be208747d8a0

+++ lib/content.functions.php
@@ -273,19 +273,25 @@ class Smarty_CMS extends Smarty {
 		$fname = '';
 		if (is_writable($config["previews_path"]))
 		{
-			$fname = $config["previews_path"] . "/" . $tpl_name;
+			$fname = cms_join_path($config["previews_path"] , $tpl_name);
 		}
 		else
 		{
-			$fname = TMP_CACHE_LOCATION . '/' . $tpl_name;
+			$fname = cms_join_path(TMP_CACHE_LOCATION , $tpl_name);
 		}
-		$handle = fopen($fname, "r");
-		$data = unserialize(fread($handle, filesize($fname)));
-		fclose($handle);
-		unlink($fname);
-
-		$tpl_source = $data["template"];
-
+		if (true == file_exists($fname))
+		{
+			$handle = fopen($fname, "r");
+			$data = unserialize(fread($handle, filesize($fname)));
+			fclose($handle);
+			unlink($fname);
+	
+			$tpl_source = $data["template"];
+		}
+		else
+		{
+			$tpl_source = 'Error: Cache file: '.$tpl_name.' does not exist.';
+		}
 		#Perform the content template callback
 		reset($gCms->modules);
 		while (list($key) = each($gCms->modules))
